Rewriting DatabaseManager to be an inheritable class
