+++ Benchmarks/Benchmarks.csproj
@@ -27,4 +27,10 @@
   <ItemGroup>
     <PackageReference Update="Nerdbank.GitVersioning" Version="3.1.91" />
   </ItemGroup>
+  <ItemGroup>
+    <Compile Remove="OpenTransactionTest.cs" />
+    <Compile Remove="QueryTests.cs" />
+    <Compile Remove="SetupTimeTest.cs" />
+    <Compile Remove="InsertTestsWithoutTransactions.cs" />
+  </ItemGroup>
 </Project>
