+++ Tests/Tests.csproj
@@ -58,4 +58,8 @@
   <ItemGroup>
     <ProjectReference Include="..\Lib\Lib.csproj" />
   </ItemGroup>
+  <ItemGroup>
+    <Compile Remove="InDatabaseComparatorTests.cs" />
+    <Compile Remove="DatabaseManagerTests.cs" />
+  </ItemGroup>
 </Project>
